feat: refactor code #168781818
