+++ src/validation/checkRoles.js
@@ -2,7 +2,7 @@ import validator from '../utils/validator';
 import sendResult from '../utils/sendResult';
 import db from '../database/models/index';
 
-export const checkRole = async (req, res, next) => {
+const checkRole = async (req, res, next) => {
   try {
     new validator({ email: req }).req().email();
     new validator({ new_role: req }).req().min(2).alpha();
@@ -14,15 +14,9 @@ export const checkRole = async (req, res, next) => {
 
   if (!checkIfExist) return sendResult(res, 400, 'Role Value does not exist');
 
+  req.new_role_id = checkIfExist.id;
+
   return next();
 };
 
-export const newRole = async (req, res, next) => {
-  try {
-    new validator({ role_name: req }).req().min(4).str();
-    new validator({ role_value: req }).req().min(4).str();
-  } catch (error) {
-    return sendResult(res, 400, error.message);
-  }
-  next();
-};
+export default checkRole;
